Fix warning that NUll was not a valid integer value
Upon insertion, warnings would show stating that 'NULL' was not a valid
intger value for rev_parent_id. Fixed by changing the default value in
the schema from 'NULL' to None, effectively casuing the database to
populate the table with NULLS instead of inserting them. Similarly, NULL
values in the revision.csv are now replaced with None